replaced celery with django-q, enabled tasks feature
