+++ .dockerignore
@@ -1,0 +1,6 @@
+venv/
+*.pyc
+__pycache__/
+coldfront.db
+coldfront.egg-info
+.gitignore
