+++ coldfront/__init__.py
@@ -5,8 +5,8 @@ import sys
 
 __version__ = '1.1.6'
 VERSION = __version__
-__all__ = ('celery_app',)
-from .celery import app as celery_app
+
+
 
 
 def manage():
