+++ coldfront/config/settings.py
@@ -50,10 +50,3 @@ for lc in local_configs:
     coldfront_configs.append(optional(lc))
 
 include(*coldfront_configs)
-
-CELERY_BROKER_URL = 'redis://localhost:6379/0' 
-CELERY_RESULT_BACKEND = 'redis://localhost:6379/0'
-CELERY_ACCEPT_CONTENT = ['json']
-CELERY_TASK_SERIALIZER = 'json'
-CELERY_RESULT_SERIALIZER = 'json'
-CELERY_TIMEZONE = 'UTC'
