+++ coldfront/core/utils/management/commands/add_scheduled_tasks.py
@@ -23,3 +23,7 @@ class Command(BaseCommand):
         schedule('coldfront.core.allocation.tasks.send_expiry_emails',
                  schedule_type=Schedule.DAILY,
                  next_run=date)
+
+        schedule('coldfront.core.allocation.tasks.update_allocations_usage',
+                 schedule_type=Schedule.HOURLY,
+                 next_run=date)
