+++ requirements.txt
@@ -2,14 +2,12 @@ arrow==1.3.0
 asgiref==3.7.2
 bibtexparser==1.4.1
 blessed==1.20.0
-celery
 certifi==2024.2.2
 chardet==5.2.0
 charset-normalizer==3.3.2
 Django==4.2.11
 django-crispy-forms==2.1
 crispy-bootstrap4==2024.1
-django-celery-beat
 django-environ==0.11.2
 django-model-utils==4.4.0
 django-picklefield==3.1
@@ -25,6 +23,7 @@ Faker==24.1.0
 fontawesome-free==5.15.4
 FormEncode==2.1.0
 future==1.0.0
+gunicorn
 humanize==4.9.0
 idna==3.6
 pyparsing==3.1.2
